# consul: improve consul service discovery