@brian-brazil 2018-02-08
I'm wondering if we could get away with just RefreshInterval, the other settings will be less useful/counter-productive once #2191 is in.