@iksaif 2018-02-08
Update the patchs

> The config documentation needs updating.

This should be done now. Please tell me if this is still unclear.

> Mixing in the core change plus several optimisations on top of it makes this a bit challenging to review.

Sorry for that :/

> I'm wondering if we could get away with just RefreshInterval, the other settings will be less useful/counter-productive once #2191 is in.

I guess that would still be an improvement. But I believe exposing the consul filtering capabilities has a real value because there is a non-significant overhead when using relabeling to do the filtering (as seen on the table).