@iksaif 2018-02-12
We deployed this patch on our pre-production cluster (~1k services), here are the results:

![consul](https://user-images.githubusercontent.com/1032963/36105886-0a5d5d12-1016-11e8-9a14-b4a5d64d3eb2.png)
![consul-2](https://user-images.githubusercontent.com/1032963/36105878-06bb8e40-1016-11e8-8e01-726e8138dfb8.png)

Considering that Prometheus is far from being the only user of Consul, this is pretty impressive :)