@cstyan 2018-02-13
lgtm, @brian-brazil ? I think we can merge before #2191 since I don't know of anyone actively working on the fix for this