@brian-brazil 2018-02-13
> I guess that would still be an improvement. But I believe exposing the consul filtering capabilities (currently tag + node-meta, soon service-meta) has a real value because there is a non-significant overhead when using relabeling to do the filtering (as seen on the table).

It'd still be interested in knowing what just the delay would give us. The other settings are going to cause support hassle, and reduce the effectiveness of #2191.

> I think we can merge before #2191 since I don't know of anyone actively working on the fix for this

That was never blocking this.