@iksaif 2018-02-13
> It'd still be interested in knowing what just the delay and allowstale would give us. 

Here is the table with the new requested column (`Delay`) with `refresh_interval: 10s`

| Resource | Empty | Before | After | After + delay | Delay |
| -------- |:-----:|:------:|:-----:|:-------------:|
|/service-discovery size|5K|85MiB|27k|27k|27k|**92MiB**|
|`go_memstats_heap_objects`|100k|**1M**|120k|110k|**1M**|
|`go_memstats_heap_alloc_bytes`|24MB|**150MB**|28MB|27MB|**140MiB**|
|`rate(go_memstats_alloc_bytes_total[5m])`|0.2MB/s|**28MB/s**|2MB/s|0.3MB/s|**23MB/s**|
|`rate(process_cpu_seconds_total[5m])`|0.1%|**15%**|2%|0.01%|**15%**|
|`process_open_fds`|16|*1236*|22|22|18|
|`rate(prometheus_sd_consul_rpc_duration_seconds_count{call="services"}[5m])`|~0|1|1|**0.03**|0.1|
|`rate(prometheus_sd_consul_rpc_duration_seconds_count{call="service"}[5m])`|0.1|**80**|0.5|0.5|**90**|
|`prometheus_target_sync_length_seconds{quantile="0.9",scrape_job="observability-by-tag"}`|N/A|200ms|0.2ms|0.2ms|180ms|
|Network bandwidth|~10kbps|**~2.8Mbps**|**~1.6Mbps**|~10kbps|**~2Mbps**|

Something weird, process_open_fds is pretty low, but I *did* have the `socket: too many open files` before I played with ulimit. This is probably because this happens only on startup and after the the `time.Sleep()` is preventing all the things to happen at the exact same time.. maybe (or it's a gauge, so we can't know).

The other results are what would be expected: filtering only with relabel_config means that all the information needs to be fetched by the client (prometheus) *before* filtering, so this is less efficient than doing it server side (even if with the patch tag filtering is *partly* client-side, but still more efficient).