@brian-brazil 2018-02-14
That result doesn't seem quite consistent with the rest, for example service calls shouldn't have gone up. Can you double check you performed the test correctly?

@brancz This PR contains major changes to Consul SD. Do you think we should hold it for after 2.2 for the sake of stabilising things?