@iksaif 2018-02-14
You are right, this seems counter intuitive, but let's look at the numbers

I have 1235 services. Using relabeling I'll watch each of them.
With a refresh_interval or 10s I expect to have 123 calls per second at maximum.
With a static consul cluster, with a `WaitTime` of 30s, this should be at minimum 41 calls per second.

This means that the expected numbers are:
* **A** `refresh_interval: 0s`: [41, +inf[ (bounded by latency)
* **B** `refresh_interval: 10s`: [41, 123[

But you would expect A >> B. Which doesn't seem to be the case here. In facts it looks like there are *more* calls when refresh_interval is non-zero, which is confirmed by the following trace:
```
level=debug ts=2018-02-14T14:18:12.817694053Z caller=consul.go:419 component="discovery manager scrape" discovery=consul msg="Watching service" service=observability-portal-web tag=
level=debug ts=2018-02-14T14:18:12.825106702Z caller=consul.go:428 component="discovery manager scrape" discovery=consul msg="Watching service !!done!!" service=observability-portal-web tag= meta.LastIndex=849509831 LastIndex=849509470
level=debug ts=2018-02-14T14:18:22.825509212Z caller=consul.go:419 component="discovery manager scrape" discovery=consul msg="Watching service" service=observability-portal-web tag=
level=debug ts=2018-02-14T14:18:22.826759492Z caller=consul.go:428 component="discovery manager scrape" discovery=consul msg="Watching service !!done!!" service=observability-portal-web tag= meta.LastIndex=849510183 LastIndex=849509831
level=debug ts=2018-02-14T14:18:32.827078498Z caller=consul.go:419 component="discovery manager scrape" discovery=consul msg="Watching service" service=observability-portal-web tag=
level=debug ts=2018-02-14T14:18:32.828510762Z caller=consul.go:428 component="discovery manager scrape" discovery=consul msg="Watching service !!done!!" service=observability-portal-web tag= meta.LastIndex=849510481 LastIndex=849510183
level=debug ts=2018-02-14T14:18:42.828818192Z caller=consul.go:419 component="discovery manager scrape" discovery=consul msg="Watching service" service=observability-portal-web tag=
level=debug ts=2018-02-14T14:18:42.829889865Z caller=consul.go:428 component="discovery manager scrape" discovery=consul msg="Watching service !!done!!" service=observability-portal-web tag= meta.LastIndex=849510787 LastIndex=849510481
level=debug ts=2018-02-14T14:18:52.830123327Z caller=consul.go:419 component="discovery manager scrape" discovery=consul msg="Watching service" service=observability-portal-web tag=
level=debug ts=2018-02-14T14:18:52.831354479Z caller=consul.go:428 component="discovery manager scrape" discovery=consul msg="Watching service !!done!!" service=observability-portal-web tag= meta.LastIndex=849511144 LastIndex=849510787
level=debug ts=2018-02-14T14:19:02.831627833Z caller=consul.go:419 component="discovery manager scrape" discovery=consul msg="Watching service" service=observability-portal-web tag=
level=debug ts=2018-02-14T14:19:02.832902286Z caller=consul.go:428 component="discovery manager scrape" discovery=consul msg="Watching service !!done!!" service=observability-portal-web tag= meta.LastIndex=849511497 LastIndex=849511144
```

This happens as soon as the `refresh_interval` is bigger than 2sec, meaning that it effectively becomes a regular refresh interval. I'm trying to dig into consul's code to understand what's happening.